# ApiDoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Call Center",
        "description": "Daftar Api Call Center",
        "version": "1.0.0"
    },
...
    "servers": [
        {
            "url": "http://192.168.3.98:8008",
            "description": "Server Staging"
        },
        {
            "url": "http://localhost:8008",
            "description": "Server Local"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Insert Data Users",
                "description": "",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "username",
                                    "password",
                                    "user_level_id",
                                    "department_id",
                                    "fullname"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "user_level_id": {
                                        "type": "integer"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "user_activity_id": {
                                        "type": "integer"
                                    },
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "middlename": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "picture": {
                                        "type": "string"
                                    },
                                    "gender_id": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get By Id Users",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Users to get."
                    }
                ],
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update By Id Users",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Users to get."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "user_level_id": {
                                        "type": "integer"
                                    },
                                    "department_id": {
                                        "type": "integer"
                                    },
                                    "user_activity_id": {
                                        "type": "integer"
                                    },
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "middlename": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "picture": {
                                        "type": "string"
                                    },
                                    "gender_id": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
        "/quality_assurance": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssurance"
                ],
                "summary": "Get All QualityAssurance",
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssurance"
                ],
                "summary": "Insert QualityAssurance",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "is_active"
                                ],
                                "properties": {
                                    "sort": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "info": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
        "/quality_assurance/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssurance"
                ],
                "summary": "Get By Id QualityAssurance",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the QualityAssurance to get."
                    }
                ],
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssurance"
                ],
                "summary": "Update By Id QualityAssurance",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the QualityAssurance to update."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "sort": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "info": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
        "/quality_assurance_users": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssuranceUsers"
                ],
                "summary": "Get All QualityAssuranceUsers",
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssuranceUsers"
                ],
                "summary": "Insert QualityAssuranceUsers",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "quality_assurance_id",
                                    "user_id",
                                    "call_id"
                                ],
                                "properties": {
                                    "quality_assurance_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "call_id": {
                                        "type": "integer"
                                    },
                                    "quality_assurance_date": {
                                        "type": "date"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
        "/quality_assurance_users/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssuranceUsers"
                ],
                "summary": "Get By Id QualityAssuranceUsers",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the QualityAssuranceDetails to get."
                    }
                ],
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "QualityAssuranceUsers"
                ],
                "summary": "Update By Id QualityAssuranceUsers",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the QualityAssuranceUsers to update."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "quality_assurance_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "call_id": {
                                        "type": "integer"
                                    },
                                    "quality_assurance_date": {
                                        "type": "date"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}